+++ src/DirectAdmin/DirectAdmin.php
@@ -8,14 +8,14 @@
  * file that was distributed with this source code.
  */
 
-namespace Omines\DirectAdmin;
+namespace Mvdgeijn\DirectAdmin;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\TransferException;
-use Omines\DirectAdmin\Context\AdminContext;
-use Omines\DirectAdmin\Context\ResellerContext;
-use Omines\DirectAdmin\Context\UserContext;
-use Omines\DirectAdmin\Utility\Conversion;
+use Mvdgeijn\DirectAdmin\Context\AdminContext;
+use Mvdgeijn\DirectAdmin\Context\ResellerContext;
+use Mvdgeijn\DirectAdmin\Context\UserContext;
+use Mvdgeijn\DirectAdmin\Utility\Conversion;
 
 /**
  * DirectAdmin API main class, encapsulating a specific account connection to a single server.