+++ src/DirectAdmin/Objects/DomainObject.php
@@ -8,7 +8,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Omines\DirectAdmin\Objects;
+namespace Mvdgeijn\DirectAdmin\Objects;
 
 /**
  * Encapsulates a domain-bound object.